# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: KMX EDC Console BFF API
  version: "1.0.0"
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    데이터스페이스 커넥터(EDC) 운영 콘솔의 BFF API.

    ## 인증 (Cookie 세션)
    - 로그인 성공 시 서버가 두 쿠키를 발급한다:
      - `kmx_token` — httpOnly JWT 세션(클라 JS 접근 불가), `SameSite=Strict`, prod 에서 `Secure`.
      - `kmx_csrf` — JS 가독 CSRF 이중제출 토큰.
    - 이후 모든 요청은 쿠키가 자동 첨부된다. 프로그램 클라이언트는
      `Authorization: Bearer <jwt>` 폴백도 사용할 수 있다.

    ## CSRF (double-submit)
    - 상태 변경 메서드(POST/PUT/PATCH/DELETE)는 `X-CSRF-Token` 헤더에 `kmx_csrf`
      쿠키 값을 되돌려 실어야 한다. 불일치 시 403.
    - 예외: `POST /api/auth/login` (자격증명으로 정당성이 검증되고 부작용이 세션 발급뿐).

    ## RBAC
    - 역할: `admin` > `operator` > `viewer`. 쓰기(생성/수정/삭제)는 operator 이상,
      플랫폼/설정은 admin. 권한 부족 시 403 `forbidden`.

    ## 에러 봉투
    - 모든 오류 응답은 `{ "error": string }` 를 기본으로 하며, EDC/DTR 프록시 오류는
      `source`("edc"|"dtr")와, 4xx 인 경우 `actionable: true` 를 덧붙일 수 있다.

    ## 멀티테넌시
    - 세션 JWT 의 `tenantId` 로 모든 리소스가 스코프된다. 커넥터 하위 경로는
      소유권 검사(requireConnectorOwnership)를 통과해야 한다(교차테넌트 접근 시 404).

    ## 페이지네이션
    - 목록 엔드포인트는 전체 배열을 반환하고 페이지네이션은 클라이언트에서 수행한다
      (usePagination + DataTablePagination). 서버측 page/limit 쿼리는 사용하지 않는다.
servers:
  - url: /
    description: 동일 오리진(콘솔이 BFF 와 같은 호스트에서 서빙됨)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/connectors:
    get:
      tags: [connectors]
      summary: 커넥터 목록 (테넌트 스코프, 상태 포함)
      responses:
        "200":
          description: 커넥터 배열
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Connector" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [connectors]
      summary: 커넥터 등록
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectorInput" }
      responses:
        "201":
          description: 생성됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "400": { $ref: "#/components/responses/ConnectorInputError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/connectors/{id}:
    parameters:
      - $ref: "#/components/parameters/ConnectorId"
    put:
      tags: [connectors]
      summary: 커넥터 수정
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectorInput" }
      responses:
        "200":
          description: 수정됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "400": { $ref: "#/components/responses/ConnectorInputError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [connectors]
      summary: 커넥터 삭제
      responses:
        "204": { description: 삭제됨 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/connectors/{id}/bpn-groups/bpn/{bpn}:
    parameters:
      - $ref: "#/components/parameters/ConnectorId"
      - name: bpn
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [bpn-groups]
      summary: 특정 BPN 이 속한 그룹 목록
      responses:
        "200":
          description: BPN 의 그룹 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  bpn: { type: string }
                  groups:
                    type: array
                    items: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/connectors/{id}/bpn-groups/{bpn}:
    parameters:
      - $ref: "#/components/parameters/ConnectorId"
      - name: bpn
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [bpn-groups]
      summary: BPN 의 그룹 매핑 삭제
      responses:
        "204": { description: 삭제됨 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── stats ──────────────────────────────────────────────────────────────────
...
  /api/notifications:
    get:
      tags: [notifications]
      summary: 알림 목록
      responses:
        "200":
          description: 알림 배열
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [notifications]
      summary: 알림 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Notification" }
      responses:
        "201":
          description: 생성됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CsrfForbidden" }
    delete:
      tags: [notifications]
      summary: 전체 알림 삭제
      responses:
        "204": { description: 삭제됨 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CsrfForbidden" }
...
  /api/audit:
    get:
      tags: [audit]
      summary: 감사 로그 조회(테넌트 스코프)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 500, maximum: 2000 }
        - name: days
          in: query
          description: 서버측 기간 필터(1~90). 미지정 시 전체 기간.
          schema: { type: integer, minimum: 1, maximum: 90 }
      responses:
        "200": { $ref: "#/components/responses/EdcObjectArray" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ── field-history ────────────────────────────────────────────────────────────
  /api/field-history:
    get:
      tags: [field-history]
      summary: 필드 변경 이력 조회
      responses:
        "200": { $ref: "#/components/responses/EdcObjectArray" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [field-history]
      summary: 필드 변경 이력 기록
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EdcObject" }
      responses:
        "201": { $ref: "#/components/responses/EdcObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/CsrfForbidden" }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: kmx_token
      description: |
        httpOnly JWT 세션 쿠키. 로그인 시 발급된다. 상태 변경 요청은 추가로
        `X-CSRF-Token` 헤더에 `kmx_csrf` 쿠키 값을 실어야 한다(double-submit).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 프로그램 클라이언트용 Authorization 헤더 폴백(쿠키 대체).
  parameters:
    ConnectorId:
      name: id
      in: path
      required: true
      description: 커넥터 ID(테넌트 소유권 검사 대상)
      schema: { type: string }
...
  schemas:
    ValidationError:
      type: object
      required: [error, issues]
      description: zod 입력 검증 실패 봉투(로그인·비밀번호변경·커넥터 등록/수정/연결테스트).
      properties:
        error: { type: string, const: validation }
        issues:
          type: array
          items:
            type: object
            properties:
              path: { type: string, description: "필드 경로(예: tenantId)" }
              message: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 사람이 읽을 오류 메시지(민감정보는 서버에서 일반화됨)
        source:
          type: string
          enum: [edc, dtr]
          description: 프록시 상류 오류의 출처(있을 때만)
        actionable:
          type: boolean
          description: 4xx 상류 오류에서 상세를 그대로 노출해도 되는지(admin/operator 대상)
...
    Connector:
      type: object
      description: 등록된 커넥터. status 는 헬스체크로 파생되어 목록에 병합된다.
      properties:
        id: { type: string }
        name: { type: string }
        managementUrl: { type: string, format: uri }
        protocolUrl: { type: string, format: uri }
        bpn: { type: string }
        status:
          type: string
          enum: [up, warn, down]
          description: 4개 프로브(관리 API·health·DSP·DB) 중 ok 수로 판정
        tenantId: { type: string }
      additionalProperties: true
    ConnectorInput:
      type: object
      required: [name, managementUrl]
      properties:
        name: { type: string }
        managementUrl: { type: string, format: uri }
        protocolUrl: { type: string, format: uri }
        apiKey:
          type: string
          description: 관리 API 키. 저장 시 AES-256-GCM 으로 암호화된다(응답에는 미포함).
        bpn: { type: string }
      additionalProperties: true
    Notification:
      type: object
      description: 콘솔 알림(폴러 생성). source 는 알림 토글(NotifyPrefs)과 대응한다.
      properties:
        id: { type: string }
        type: { type: string, enum: [info, warning, error, success] }
        source:
          type: string
          enum: [system, negotiation, transfer, edr, vc]
        title: { type: string }
        message: { type: string }
        link: { type: ["string", "null"] }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
      additionalProperties: true
...
    EdcObject:
      type: object
      description: |
        EDC/DTR 상류가 반환하는 JSON-LD 오브젝트를 그대로 전달한 것. 스키마를 고정하지
        않는다(@context/@id/@type 등 JSON-LD 키 포함 가능). 정확한 형태는 EDC 0.17.0
        Management API 계약을 참조.
      additionalProperties: true
  responses:
    EdcObject:
      description: EDC/DTR 상류 오브젝트(pass-through)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/EdcObject" }
    EdcObjectArray:
      description: EDC/DTR 상류 오브젝트 배열(pass-through, 클라이언트 페이지네이션)
      content:
        application/json:
          schema:
            type: array
            items: { $ref: "#/components/schemas/EdcObject" }
...
    ConnectorInputError:
      description: 입력 검증(zod) 실패 또는 SSRF 거부(사설/메타데이터 URL)
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ValidationError"
              - $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: 미인증(토큰 없음/무효/만료/무효화) 또는 세션 만료
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 인증됐으나 권한 부족(RBAC) 또는 CSRF 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    CsrfForbidden:
      description: CSRF 토큰 불일치 또는 권한 부족
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 리소스 없음 또는 교차테넌트 접근(소유권 검사 실패)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }